+++ crates/analyzer/src/handlers/check_instance.rs
@@ -43,43 +43,44 @@ impl<'a> VerylGrammarTrait for CheckInstance<'a> {
                 }
             }
 
-            let symbol = symbol_table::resolve(arg.identifier0.as_ref());
-            let name = arg.identifier0.identifier_token.text();
-            if let Some(symbol) = symbol {
-                match symbol.kind {
-                    SymbolKind::Module(ref x) => {
-                        for port in &x.ports {
-                            if !connected_ports.contains(&port.name) {
-                                let port = resource_table::get_str_value(port.name).unwrap();
-                                self.errors.push(AnalyzerError::missing_port(
-                                    &name,
-                                    &port,
-                                    self.text,
-                                    &arg.identifier.identifier_token,
-                                ));
+            if let Ok(symbol) = symbol_table::resolve(arg.identifier0.as_ref()) {
+                let name = arg.identifier0.identifier_token.text();
+                if let Some(symbol) = symbol.found {
+                    match symbol.kind {
+                        SymbolKind::Module(ref x) => {
+                            for port in &x.ports {
+                                if !connected_ports.contains(&port.name) {
+                                    let port = resource_table::get_str_value(port.name).unwrap();
+                                    self.errors.push(AnalyzerError::missing_port(
+                                        &name,
+                                        &port,
+                                        self.text,
+                                        &arg.identifier.identifier_token,
+                                    ));
+                                }
+                            }
+                            for port in &connected_ports {
+                                if !x.ports.iter().any(|x| &x.name == port) {
+                                    let port = resource_table::get_str_value(*port).unwrap();
+                                    self.errors.push(AnalyzerError::unknown_port(
+                                        &name,
+                                        &port,
+                                        self.text,
+                                        &arg.identifier.identifier_token,
+                                    ));
+                                }
                             }
                         }
-                        for port in &connected_ports {
-                            if !x.ports.iter().any(|x| &x.name == port) {
-                                let port = resource_table::get_str_value(*port).unwrap();
-                                self.errors.push(AnalyzerError::unknown_port(
-                                    &name,
-                                    &port,
-                                    self.text,
-                                    &arg.identifier.identifier_token,
-                                ));
-                            }
+                        SymbolKind::Interface(_) => (),
+                        _ => {
+                            self.errors.push(AnalyzerError::mismatch_type(
+                                &name,
+                                "module or interface",
+                                &symbol.kind.to_kind_name(),
+                                self.text,
+                                &arg.identifier.identifier_token,
+                            ));
                         }
-                    }
-                    SymbolKind::Interface(_) => (),
-                    _ => {
-                        self.errors.push(AnalyzerError::mismatch_type(
-                            &name,
-                            "module or interface",
-                            &symbol.kind.to_kind_name(),
-                            self.text,
-                            &arg.identifier.identifier_token,
-                        ));
                     }
                 }
             }