+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased](https://github.com/dalance/veryl/compare/v0.2.0...Unreleased) - ReleaseDate
 
+* [Added] uknown member check [#84](https://github.com/dalance/veryl/issues/84)
+
 ## [v0.2.0](https://github.com/dalance/veryl/compare/v0.1.14...v0.2.0) - 2023-01-16
 
 * [Changed] modport separator from `.` to `::` [#65](https://github.com/dalance/veryl/issues/65)